Merge

+++ app/models.py
@@ -42,9 +42,9 @@ class Media(db.Model):
         return "<Media {}>".format(self.title)
 
     # Deletes a particular piece of media
-    def delete_media(self):
-        db.session.delete(self)
-        db.session.commit()
+    #def delete_media(self):
+    #    db.session.delete(self)
+    #    db.session.commit()
 
 # Stores all polls
 class Poll(db.Model):
@@ -53,6 +53,7 @@ class Poll(db.Model):
     creator = db.Column(db.Integer, db.ForeignKey("user.id"))       # The id of the user who created the poll
     timestamp = db.Column(db.DateTime, default = datetime.utcnow)   # The time and date of creation
     active = db.Column(db.Boolean(), default = True)                # Whether or not poll can be voted on
+    keyword = db.Column(db.String(20), default = "best")            # Used to display immediate purpose of poll
     choices = db.relationship("Media", secondary = "global_polls",  # Returns all media in this poll
                                 backref = "poll")
     #backref = author. Returns creator of this poll
@@ -70,16 +71,20 @@ class Poll(db.Model):
     def add_media(self, media):
         if not self.contains(media):
             self.choices.append(media)
+            return True
+        return False
     
     # Remove media from poll(I don't know if this will be used but it's good to have)
     def remove_media(self, media):
         if self.contains(media):
             self.choices.remove(media)
+            return True
+        return False
 
     # Deletes poll
-    def delete_poll(self):
-        db.session.delete(self)
-        db.session.commit()
+    #def delete_poll(self):
+    #    db.session.delete(self)
+    #    db.session.commit()
     
     # Returns all participants
     def voters(self):
@@ -98,15 +103,15 @@ class Poll(db.Model):
             tot.append({"Media" : gp.parent_med, "GlobalScore" : tally})
         return tot
 
-    #Returns one movie poster
+    # Returns one movie poster
     def cover(self):
         if len(self.choices) == 0:
             return "img/poster.png"
         return self.choices[0].poster
 
-    #Deactivtes all current open polls
+    # Deactivates all current open polls
     def close_all():
-        for p in Poll.query.filter(Poll.active).all():
+        for p in Poll.query.filter(Poll.active == True).all():
             p.active = False
         
 # Stores all users
@@ -136,21 +141,23 @@ class User(UserMixin, db.Model):
         db.session.delete(self)
         db.session.commit()
 
-    # checks if user has already participated in a poll
+    # Checks if user has already participated in a poll
     def already_voted(self, poll):
         for up in self.votes:
-            if up.p_gp.parent_poll == poll:
-                return True
+            if not up.p_gp == None:
+                if up.p_gp.parent_poll == poll:
+                    return True
         return False
 
     # Removes previous votes(Used before redoing a poll)
     def remove_user_poll(self, poll):
         if self.already_voted(poll):
             for up in self.votes:
-                if up.p_gp.parent_poll == poll:
-                    db.session.delete(up)
+                if not up.p_gp == None:
+                    if up.p_gp.parent_poll == poll:
+                        db.session.delete(up)
 
-    #Votes on a single entry in a single poll
+    # Votes on a single entry in a single poll
     def vote_on_media(self, poll, media, score):
         for gp in poll.associates:
             if gp.m_id == media.id:
@@ -161,15 +168,18 @@ class User(UserMixin, db.Model):
     def all_polls(self):
         ap = []
         for up in self.votes:
-            if not up.p_gp.parent_poll in ap:
-                ap.append(up.p_gp.parent_poll)
+            if not up.p_gp == None:
+                if not up.p_gp.parent_poll in ap:
+                    ap.append(up.p_gp.parent_poll)
         return ap
 
+    # Returns results of poll if user has participated in it
     def poll_results(self, poll):
         pr = []
         for up in self.votes:
-            if up.p_gp.parent_poll == poll:
-                pr.append({"Media" : up.p_gp.parent_med , "Score" : up.score})
+            if not up.p_gp == None:
+                if up.p_gp.parent_poll == poll:
+                    pr.append({"Media" : up.p_gp.parent_med , "Score" : up.score})
         return pr
 
 @login.user_loader
